Split Functions

+++ srcs/utils/angles.c
@@ -6,7 +6,7 @@
 /*   By: arebelo <arebelo@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/18 18:41:56 by arebelo           #+#    #+#             */
-/*   Updated: 2023/04/19 15:21:52 by arebelo          ###   ########.fr       */
+/*   Updated: 2023/04/20 18:15:59 by arebelo          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,14 @@
 /**
  * Returns the angle (given in degrees) in radians 
 */
-float	deg_to_rad(int d)
+float	deg_to_rad(float d)
 { 
 	return d * M_PI/ 180.0;
 }
 
+/**
+ * Corrects angle to always be between 0 and 359 (inclusive)
+*/
 float	angle_check(float ang)
 {
 	if (ang > 359)
